상세페이지 작업중

+++ app/src/main/java/com/example/newbookshelf/data/model/detail/review/UpdateBookReviewData.kt
@@ -2,5 +2,5 @@ package com.example.newbookshelf.data.model.detail.review
 
 data class UpdateBookReviewData(
     val comment_content: String,
-    val comment_rate: Int
+    val comment_rate: Double
 )
